Move activity to home view
